Fixed shortcuts, impossible parts
Fixed a major shortcut in [UpDownUpDown] where the player would spawn on top of the level on a higher refresh rate, skipping the first ascension part. Also fixed a shortcut on [Baleful Bridge] that allowed the player to jump out of the starting chute.

Fixed a section of [Baleful Bridge] that was previously impossible without glitching on lower refresh rates.